+++ jvm/src/main/scala/no/skytteren/scalatime/package.scala
@@ -5,14 +5,14 @@ import java.time.LocalDateTime
 package object scalatime {
 
   implicit class TimePimp(from: Time) {
-    def toJavaLocalTime: java.time.LocalTime = java.time.LocalTime.of(from.hour.value, from.minute.value, from.second.value, from.millisecond.value * 1000000)
+    def toJavaLocalTime: java.time.LocalTime = java.time.LocalTime.of(from.hour.value.toInt, from.minute.value, from.second.value.toInt, from.millisecond.value.toInt * 1000000)
   }
   implicit class TimeObjectPimp(time: Time.type) {
     def now(): Time = java.time.LocalTime.now().toTime
   }
 
   implicit class DatePimp(date: Date){
-    def toJavaLocalDate: java.time.LocalDate = java.time.LocalDate.of(date.year.value.toInt, date.month.value, date.dayOfMonth.value)
+    def toJavaLocalDate: java.time.LocalDate = java.time.LocalDate.of(date.year.value.toInt, date.month.value.toInt, date.dayOfMonth.value.toInt)
   }
   implicit class DateObjectPimp(date: Date.type) {
     def now(): Date = {
@@ -24,8 +24,8 @@ package object scalatime {
   implicit class DateTimePimp(datetime: DateTime) {
     def toJavaLocalDateTime: LocalDateTime = {
       LocalDateTime.of(
-        datetime.year.value.toInt, datetime.month.value, datetime.dayOfMonth.value,
-        datetime.hour.value, datetime.minute.value, datetime.second.value, datetime.millisecond.value * 1000000
+        datetime.year.value, datetime.month.value.toInt, datetime.dayOfMonth.value.toInt,
+        datetime.hour.value.toInt, datetime.minute.value, datetime.second.value, datetime.millisecond.value * 1000000
       )
     }
   }
